Refactoring to make it easier for extension
